% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discrepancy_completed.R
\name{get_discrepancy_completed}
\alias{get_discrepancy_completed}
\title{Get discrepancy measure computed using the completed data and the replicated completed data.}
\usage{
get_discrepancy_completed(K, C, priorPik, bSub, betaObs, Sigma, Y, XRe,
  XObs, subjectIDY)
}
\value{
Discrepancy measure computed using completed data and the replicated completed data.
}
\description{
For a discrepancy measure for the model of \code{Y1,\dots,YJ}, the multivariate mean squared error is used. References for using a discrepancy measure with completed and replicated completed data include: 1. Xu, D., Chatterjee, A. and Daniels, M. (2016) A Note on Posterior Predictive Checks to Assess Model Fit for Incomplete Data. Statistics in Medicine, 35, 5029-5039. 2. Gelman, A., Mechelen, I. V., Verbeke, G., Heitjan, D.F. and Meulders, M. (2005) Multiple Imputation for Model Checking: Completed-Data Plots with Missing and Latent Data. Biometrics, 61, 74-85.
}
