% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discrepancy_plot.R
\name{get_discrepancy_plot}
\alias{get_discrepancy_plot}
\title{Get discrepancy plot using completed and replicated completed data.}
\usage{
get_discrepancy_plot(discrepancySamples)
}
\arguments{
\item{discrepancySamples}{Discrepancy samples written to \code{store_T_completed.txt} in \code{MVNYBinaryMiss}, when \code{sims = FALSE}.}
}
\value{
A scatter plot of the discrepancy measure calculated using the completed data versus the replicated completed data. The plot is annotated with the posterior predictive p-value.
}
\description{
Overall adequacy of the fitted model can be examine using Bayesian posterior predictive p-values. At each MCMC iteration, a discrepancy measure is computed using the completed and replicated completed data. The Bayesian predictive p-value denotes the probability that the discrepancy measure under the replicated data is greater than that under the observed data. A p-value near 0.5 indicates adequate model fit, while a p-value outside the range of 0.05 and 0.95 is considered to suggest a lack of model fit.

The discrepancy measure used in this function is a multivariate mean squared error, computed as in:
Daniels, M.J. and Hogan, J.W. (2008) Missing Data in Longitudinal Studies: Strategies for Bayesian Modeling and Sensitivity Analysis. Chapman and Hall/CRC, Boca Raton, p. 67.

A reference for using the completed data versus replicated completed data is in:
Gelman, A., Mechelen, I. V., Verbeke, G., Heitjan, D.F. and Meulders, M. (2005) Multiple Imputation for Model Checking: Completed-Data Plots with Missing and Latent Data. Biometrics, 61, 74-85.
}
