% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ESS.R
\name{get_ESS}
\alias{get_ESS}
\title{Get the effective sample size for the BIC calculation.}
\usage{
get_ESS(priorPik, Psi, Sigma, Y, XRe, subjectIDY)
}
\value{
Effective sample size at the current iteration.
}
\description{
This calculation of the effective sample size is used in BIC2. The reference for the approach is in Jones, R. H. (2011). Bayesian information criterion for longitudinal and clustered data. Statistics in Medicine, 30(25), 3050-3056. I extended the approach to the setting of multiple longitudinal health outcomes.
}
