% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_checks.R
\name{get_model_checks}
\alias{get_model_checks}
\title{Get posterior predictive checks including replicates from the joint model and a discrepancy measure computed using the completed data and the replicated completed data.}
\usage{
get_model_checks(
  K,
  C,
  priorPik,
  bSub,
  betaObs,
  betaSub,
  Psi,
  Sigma,
  Y,
  XRe,
  XObs,
  XSub,
  subjectIDY,
  Q,
  phi,
  Omega,
  D,
  URe,
  UObs,
  subjectID,
  L,
  lambda,
  Theta,
  M,
  VRe,
  VObs,
  subjectIDM,
  Mvec,
  modelVisit,
  modelResponse
)
}
\value{
Draws of replicated longitudinal outcomes, visit process, and response process. Discrepancy measure computed using completed data and the replicated completed data.
}
\description{
This function generates posterior predictive draws of \code{Y1,\dots,YJ}, \code{D}, and \code{M1,\dots,MJ} based on redrawing all of the latent variables, including latent class memberships and random effects from each of the submodels for the longitudinal outcomes, visit process, and response process given a clinic visit. The posterior predictive draws are written to file. In addition, for posterior predictive checking based on the completed and replicated completed \code{Y1,\dots,YJ}, a discrepancy measure \code{T}, namely, the multivariate mean squared error, is calculated. References for using a discrepancy measure with completed and replicated completed data include: 1. Xu, D., Chatterjee, A. and Daniels, M. (2016) A Note on Posterior Predictive Checks to Assess Model Fit for Incomplete Data. Statistics in Medicine, 35, 5029-5039. 2. Gelman, A., Mechelen, I. V., Verbeke, G., Heitjan, D.F. and Meulders, M. (2005) Multiple Imputation for Model Checking: Completed-Data Plots with Missing and Latent Data. Biometrics, 61, 74-85.
}
