% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed_lik_sub.R
\name{get_observed_lik_sub}
\alias{get_observed_lik_sub}
\title{Get the likelihood after integrating out latent class membership and the random effects.}
\usage{
get_observed_lik_sub(priorPik, betaObs, betaSub, Sigma, Psi, Y, XRe, XObs,
  XSub, subjectIDY, phi, Omega, D, URe, UObs, subjectID, lambda, Theta, M,
  VRe, VObs, subjectIDM, Mvec, modelVisit, modelResponse,
  monteCarlo = FALSE)
}
\value{
Observed data likelihood contribution of each of \code{n} subjects at the current iteration.
}
\description{
The likelihood after integrating out latent class membership and random effects is used in the calculation of DIC3. The DIC3 is the modified DIC proposed for the latent variable setting. The reference for the DIC3 is in Celeux, G., Forbes, F., Robert, C. P., & Titterington, D. M. (2006). Deviance Information Criteria for Missing Data Models. Bayesian Analysis, 1(4), 651-674. This likelihood is also used the calculation of BIC1 and BIC2. When \code{modelVisit = TRUE} or \code{modelResponse = TRUE}, numerical integration is used to estimate the marginal likelihoods. An option to estimate these densities using a Monte Carlo approach is available, but the option is turned off \code{monteCarlo = FALSE} because of long computation time.
}
