% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_pred.R
\name{get_post_pred}
\alias{get_post_pred}
\title{Get posterior predictive draws of the completed data at each iteration.}
\usage{
get_post_pred(J, priorPik, bSub, betaObs, Sigma, XRe, XObs, subjectIDY)
}
\value{
Draws from the posterior predictive distribution at the current iteration.
}
\description{
The reference for the approach is from algorithm 3.7 on page 91 of Fruhwirth-Schnatter, S. (2006) Finite Mixture and Markov Switching Models. Springer Science & Business Media, New York. The latent classes are redrawn using the probabilities of latent class membership obtained from the latent class membership model.
}
