% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Simulate data for \code{MVNYBinaryMiss}.}
\usage{
simdat(K, J, data, regf, delta, betaObs, betaSub, Psi, Sigma, phi, Omega,
  responseMiss, lambda, Theta, Mvec, full)
}
\arguments{
\item{K}{A scalar for the assumed number of latent classes in the GMMs.}

\item{J}{A scalar for the number of longitudinal outcomes. \code{J} must be greater than 1.}

\item{data}{A data.frame with all analysis variables, except for the outcomes that will be generated, including \code{Y1,\dots,YJ} for the longitudinal health outcomes; \code{D} for the visit process; and \code{M1,\dots,MJ} for the response process for each health outcome given a clinic visit. The dataset must contain an integer variable for the patient ID of each longitudinal measurement called \code{subjectID}.}

\item{regf}{A list of formulas for the design matrices in each sub-model. The structure is detailed below:

To improve Gibbs sampling properties, the multivariate model for \code{Y1,\dots,YJ} is updated based on a hierarchically-centered parameterisation. As such, the named formulas for the longitudinal health outcomes model are "YRe", which is the random effects design matrix that must include at least a column of 1's for a random intercept; "YObs", which is the observation-level fixed effects design matrix containing covariates that will not be associated with random effects; and, "YSub" which contains the formula for the random effects equations using subject-level covariates. Note that the formulas "YRe", "YObs", and "YSub" must be disjoint in the hierarchically-centered parameterisation.

For the multinomial probit model of latent class membership, the named formula is "LatentClass", containing patient-level covariates that may be associated with probabilty of latent class membership.

If a model for \code{D} is included, then \code{regf} contains named formulas "DObs" and "DRe" for the fixed effects design matrix and the random effects design matrix, respectively. "DRe" contains a subset of the columns of "DObs".

If a model for any of \code{M1,\dots,MJ} is included, then \code{regf} contains named formulas "MObs" and "MRe" for the fixed effects design matrix and the random effects design matrix, respectively. "MRe" contains a subset of the columns of "MObs".

Note that \code{Y1,\dots,YJ} are assumed to have the same design matrices, and the response processes \code{M1,\dots,MJ} are assumed to have the same design matrices.}

\item{delta}{\code{K-1}-column matrix of true regression coefficients for the latent class membership model.}

\item{betaObs}{True regression coefficients for the design matrix specified by \code{regf} named formula "YObs". The object is a \code{J} element list in which each element is an \code{s} by \code{K} matrix, where \code{s} is the number of regression coefficients and \code{K} is the number of latent classes, where \code{s} is the number of regression coefficients and \code{K} is the number of latent classes.}

\item{betaSub}{True regression coefficients in the design matrix specified by \code{regf} named formula "YSub". The object is a \code{J} element list in which each element is an \code{p} by \code{q} by \code{K} array, where \code{p} is the number of regression coefficients, \code{q} is the number of random effects, and \code{K} is the number of latent classes.}

\item{Psi}{True values for the variance-covariance matrix of the subject-level random effects in the longitudinal health outcomes model. The object is a \code{J} element list in which each element is an \code{q} by \code{q} by \code{K} array, where \code{q} is the number of random effects, and \code{K} is the number of latent classes.}

\item{Sigma}{True variance-covariance matrix of the longitudinal health outcomes \code{Y1,\dots,YJ}. The object is a \code{J} by \code{J} by \code{K} array.}

\item{phi}{A \code{K}-column matrix of true regression coefficients in the design matrix specified by \code{regf} named formula "DObs".}

\item{Omega}{True variance-covariance matrix of the subject-level random effects in the visit process model. The object is an \code{q} by \code{q} by \code{K} array, where \code{q} is the number of random effects, and \code{K} is the number of latent classes.}

\item{responseMiss}{Logical. If \code{TRUE}, then response process missingness will be generated for the columns specified in \code{Mvec}.}

\item{lambda}{True regression coefficients for the design matrix specified by \code{regf} named formula "MObs". The object is a list of the same number of elements as \code{Mvec}. The order of the initial values must correspond to \code{Mvec}. If \code{responseMiss = FALSE}, then use \code{NULL}.}

\item{Theta}{True variance-covariance matrix of the subject-level random effects in the response process model. The object is a list of the same number of elements as \code{Mvec}. The order of the initial values must correspond to \code{Mvec}. Each element is a \code{q} by \code{q} by \code{K} array. If \code{responseMiss = FALSE}, then use \code{NULL}.}

\item{Mvec}{Integer-valued vector indicating which response process given a clinic visit \code{M1,\dots,MJ} will be modeled under a missing not at random mechanism. For example, if the response processes of \code{Y1} and \code{Y2}, which are \code{M1} and \code{M2}, are desired, then \code{Mvec} is \code{c(1,2)}.}

\item{full}{Logical. If \code{TRUE}, then the full data for \code{Y1,\dots,YJ} are returned, before introducing any missed visits or responses. If \code{FALSE}, the data with missed visits and missed responses are returned. In this case, the data may have fewer unique subjects than in the original \code{data} because some subjects may not have any observed clinci visits.}
}
\value{
A list with 5 elements. \code{data} returns the original data with \code{Y1,\dots,YJ} after introducing missing values; \code{YC1,\dots,YCJ} which do not have missing values; \code{D} which is a binary indicator for the response process; \code{M1,\dots,MJ} which is the response process corresponding to each longitudinal outcome; \code{PrV}, which is the probability of an observed clinic visit; and \code{PrM1,\dots,PrMJ}, which is the probability of a response given a clinic visit. \code{D} and \code{M1,\dots,MJ} are coded 1 if a visit (response) is observed, and 0 otherwise. \code{M1,\dots,MJ} will be \code{NA} when \code{D=0}. \code{C} is the true latent class membership for each subject. \code{ame} are the true population-averaged regression coefficients in the longitudinal health outcomes model. \code{weight} are the true latent class weights. \code{store_attributes} is a matrix with the attributes of the data generation.
}
\description{
This makes data generated from the shared parameter model in \code{MVNYBinaryMiss}.
}
