% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_betaLC_ProbitRE.R
\name{update_betaLC_ProbitRE}
\alias{update_betaLC_ProbitRE}
\title{Update latent class-specific regression coefficients in a probit model.}
\usage{
update_betaLC_ProbitRE(C, Z, tau, UObs, URe, prior.mu, prior.Sigma, subjectID)
}
\value{
A matrix of latent class-specific regression coefficients in which columns index the \code{K} latent classes.
}
\description{
Latent class-specific regression coefficients are updated based on a Bernoulli likelihood and probit link function. A normal prior distribution is assumed. The updating function can accomodate probit models that include a subject specific random effects.
}
