% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_betaObsLC.R
\name{update_betaObsLC}
\alias{update_betaObsLC}
\title{Update regression coefficients \code{betaObs}.

#' @description An multivariate normal prior distribution is used to update the regression coefficients \code{betaObs} corresponding to the covariates in \code{XObs}, for each outcome \code{Y1,\dots,YJ}.}
\usage{
update_betaObsLC(C, betaObs, bSub, Sigma, Y, XRe, XObs, prior.mu,
  prior.Sigma, subjectID)
}
\value{
A \code{J} element list in which each element is an \code{s} by \code{K} matrix of latent class-specific regression coefficients.
}
\description{
Update regression coefficients \code{betaObs}.

#' @description An multivariate normal prior distribution is used to update the regression coefficients \code{betaObs} corresponding to the covariates in \code{XObs}, for each outcome \code{Y1,\dots,YJ}.
}
