% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bLC_ProbitRE_fast.R
\name{update_bLC_ProbitRE_fast}
\alias{update_bLC_ProbitRE_fast}
\title{Update random effects conditional on latent class membership in a probit model.}
\usage{
update_bLC_ProbitRE_fast(C, Omega, Z, phi, UObs, URe, subjectID)
}
\value{
A stacked \code{n}\code{q} by 1 vector of random effects for each subject.
}
\description{
The random effects in a probit model are updated conditional on class membership. A normal prior distribution is assumed.
}
