% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_C_fast.R
\name{update_C_fast}
\alias{update_C_fast}
\title{Update posterior latent class assignment \code{C} for each subject.}
\usage{
update_C_fast(K, Z, C, betaObs, bSub, betaSub, Sigma, Psi, phi, tau, Omega,
  lambda, kappa, Theta, Y, XRe, XObs, XSub, D, UObs, URe, M, Mvec, VObs,
  VRe, subjectIDY, subjectID, subjectIDM, modelVisit, modelResponse)
}
\value{
A list of 2 elements. An \code{n} by 1 vector of latent class assignments called \code{C}. An \code{n} by \code{K} matrix of posterior probabilities of latent class assignment.
}
\description{
This uses Bayes' theorem to update posterior latent class assignment. This is a ratio of the likelihood contribution of each subject to class \code{k} and the prior probability of belonging in class \code{k} to the likelihood contribution marginalized over latent classes. The function accomodates the case of \code{K} equal to 2 or greater than 2.
}
