% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_delta_BinaryP.R
\name{update_delta_BinaryP}
\alias{update_delta_BinaryP}
\title{Update regression coefficients \code{delta} in a binary probit model.}
\usage{
update_delta_BinaryP(Z, delta, W, prior.mu, prior.Sigma)
}
\value{
An \code{m} by 1 vector of regression coefficients in the binary probit model.
}
\description{
This uses a normal prior distribution to update the regression coefficients in the probit model.
}
