% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_delta_MNP.R
\name{update_delta_MNP}
\alias{update_delta_MNP}
\title{Update regression coefficients \code{delta} in multinomial probit model.}
\usage{
update_delta_MNP(Z, delta, W, prior.mu, prior.Sigma)
}
\value{
An \code{m} by \code{K-1} matrix of latent class-specific regression coefficients in the multinomial probit model with \code{K=1} fixed to zero.
}
\description{
This uses a multivariate normal prior distribution to update the latent class-specific regression coefficients in the multinomial probit model.
}
