% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_OmegaLC_ProbitRE.R
\name{update_OmegaLC_ProbitRE}
\alias{update_OmegaLC_ProbitRE}
\title{Update the latent class-specific variance-covariance matrices of the subject-level random effects in a probit model.}
\usage{
update_OmegaLC_ProbitRE(C, tau, prior.scale, prior.df)
}
\value{
A \code{q} by \code{q} by \code{K} array of variance-covariance matrices for the random effects, where \code{q} is the number of random effects.
}
\description{
The variance-covariance matrices of the random effects are updated conditional on class membership using an inverse-Wishart prior distribution. This updating function is based on a probit model (Bernoulli likelihood) and a multivariate normal distribution for the random effects.
}
