% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_PsiLC.R
\name{update_PsiLC}
\alias{update_PsiLC}
\title{Update variance-covariance matrix \code{Psi} of the random effects for \code{Y1,\dots,YJ}.}
\usage{
update_PsiLC(C, betaSub, bSub, Y, XSub, prior.scale, prior.df)
}
\value{
A \code{J}-element list. Each element is latent class-specific variance-covariance matrix stored in a \code{q} by \code{q} by \code{K} array. For example, \code{Psi = list(Psi1 = array( , dim = c(q, q, K)), Psi2 = array( , dim = c(q, q, K)))}. If \code{q=1}, a scalar for the variance of the random intercept replaces the variance-covariance matrix.
}
\description{
An inverse-Wishart prior distribution is used to update the variance-covariance matrix \code{Psi} of the random effects for each outcome variable. If only a random intercept is present, then the prior collapses to an inverse gamma distribution.
}
