% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SigmaLC.R
\name{update_SigmaLC}
\alias{update_SigmaLC}
\title{Update variance-covariance matrix \code{Sigma} of the \code{Y1,\dots,YJ}.}
\usage{
update_SigmaLC(C, betaObs, bSub, Y, XRe, XObs, prior.scale, prior.df,
  subjectID)
}
\value{
A \code{J} (number of outcome variables) by \code{J} (number of outcome variables) by \code{K} (number of latent class) array of latent class-specific variance-covariance matrices.
}
\description{
An inverse-Wishart prior distribution is used to update the variance-covariance matrix \code{Sigma} of the longitudinal health outcomes \code{Y1,\dots,YJ}.
}
