% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Z_BinaryP.R
\name{update_Z_BinaryP}
\alias{update_Z_BinaryP}
\title{Update latent variable \code{Z} in a binary probit model.}
\usage{
update_Z_BinaryP(Z, C, delta, W)
}
\value{
An \code{n} by 1 vector latent variable \code{Z}.
}
\description{
This uses a truncated normal distribution to update latent variable Z. This updating function must be used with \code{K} equals 2.
}
