% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Z_MNP.R
\name{update_Z_MNP}
\alias{update_Z_MNP}
\title{Update latent variable \code{Z} in multinomial probit model.}
\usage{
update_Z_MNP(Z, C, delta, W)
}
\value{
An \code{n} by \code{K - 1} matrix of latent variable \code{Z}.
}
\description{
This uses a truncated normal distribution to update latent variable Z. This updating function must be used with \code{K} greater than 2.
}
