% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ZLC_ProbitRE.R
\name{update_ZLC_ProbitRE}
\alias{update_ZLC_ProbitRE}
\title{Update latent normal draws in a probit model.}
\usage{
update_ZLC_ProbitRE(C, phi, tau, D, UObs, URe, subjectID)
}
\value{
An observation-level vector of latent normal draws.
}
\description{
The latent normal draws for a probit model with or without a subject specific random effects are updated.
}
